refactor(giveback): split GivebackRewardReveal into scaffold + scenes
Pure code move, no behavior change. GivebackRewardReveal.tsx (1044 →
235) keeps the dispatcher + public API (RewardRevealContent,
RevealSurface, RevealDialogShell, RewardRevealDialog); revealScaffold.tsx
holds the layout primitives and revealScenes.tsx the per-reward
components.
